requirements.txt
py-aiger
python-sat[aiger,approxmc,cryptosat,pblib]
